Change license to GPL
There are a few small changes I had wanted to make to this application
since long ago but I shall only make them under the condition that it
is open sourced. As it is a fairly trivial application based on
MGRS conversion code supplied by the U.S. Government (which is public
domain) I don't expect any objections to fully releasing this formerly
closed source work.