Rewrite mgrslib in JavaScript
Replace the JavaScript calls into C code with a pure JavaScript
implementation of Convert_MGRS_To_Geodetic which is based on
transliterating the C code more or less as is. In order to minimize
the risk of transcription errors during the conversion from one
language to another, the rewrite is quite literal instead of
a refactoring using more JavaScript-friendly idioms. Nevertheless,
to help prove the correctness of the rewrite, some unit tests
are provided, in the file "runtests".

This removes the need for the custom callback into Objective C
code, making the application simpler, more portable, and able
to be used standalone in a browser.